requirements.txt
aiohttp==3.13.2
discord.py==2.6.4
python-dotenv==1.2.1
RapidFuzz==3.14.3
rich==14.2.0
pytest==9.0.1
pytest-asyncio==1.3.0
pillow==12.0.0
numpy==2.3.5
d20==1.1.2
matplotlib==3.10.7
toml==0.10.2
d20distribution @ git+https://github.com/pipieter/d20distribution@main